web-service/requirements.txt
flask
psycopg2-binary
grpcio
grpcio-tools